+++ test/source/test_gray_code.cpp
@@ -7,7 +7,7 @@
 #include <vector>
 
 TEST_CASE("Generate Gray code by BRGC_gen (odd)") {
-    int cnt = 1;
+    size_t cnt = 1;
     for ([[maybe_unused]] auto i : ecgen::BRGC_gen(5)) {
         ++cnt;
     }
@@ -15,7 +15,7 @@ TEST_CASE("Generate Gray code by BRGC_gen (odd)") {
 }
 
 TEST_CASE("Generate Gray code by BRGC_gen (even)") {
-    int cnt = 1;
+    size_t cnt = 1;
     for ([[maybe_unused]] auto i : ecgen::BRGC_gen(6)) {
         ++cnt;
     }
@@ -23,7 +23,7 @@ TEST_CASE("Generate Gray code by BRGC_gen (even)") {
 }
 
 TEST_CASE("Generate all combinations by EMK_gen") {
-    int cnt = 1;
+    size_t cnt = 1;
     for ([[maybe_unused]] auto [x, y] : ecgen::EMK_gen(5, 3)) {
         ++cnt;
     }
@@ -31,7 +31,7 @@ TEST_CASE("Generate all combinations by EMK_gen") {
 }
 
 TEST_CASE("Generate Gray code (odd)") {
-    int cnt = 0;
+    size_t cnt = 0;
     for ([[maybe_unused]] auto& l : ecgen::BRGC<std::vector<bool>>(5)) {
         ++cnt;
     }
@@ -39,15 +39,15 @@ TEST_CASE("Generate Gray code (odd)") {
 }
 
 TEST_CASE("Generate Gray code (even)") {
-    int cnt = 0;
-    for ([[maybe_unused]] auto& l : ecgen::BRGC<std::vector<int>>(6)) {
+    size_t cnt = 0;
+    for ([[maybe_unused]] auto& l : ecgen::BRGC<std::vector<size_t>>(6)) {
         ++cnt;
     }
     CHECK(cnt == 64);
 }
 
 TEST_CASE("Generate all combinations (odd)") {
-    int cnt = 0;
+    size_t cnt = 0;
     auto S = std::string("11100");
     for ([[maybe_unused]] auto& s : ecgen::EMK(5, 3, S)) {
         ++cnt;
@@ -56,10 +56,10 @@ TEST_CASE("Generate all combinations (odd)") {
 }
 
 TEST_CASE("Generate all combinations (even)") {
-    constexpr int N = 6;
-    constexpr int K = 3;
-    int cnt = 0;
-    auto lst = std::vector<int>(N, 0);
+    constexpr size_t N = 6;
+    constexpr size_t K = 3;
+    size_t cnt = 0;
+    auto lst = std::vector<size_t>(N, 0);
     std::fill_n(lst.begin(), K, 1);
     for ([[maybe_unused]] auto& l : ecgen::EMK(N, K, lst)) {
         ++cnt;
